requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
avro==1.11.2
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
construct==2.10.68
cryptography==39.0.2
dataclasses-json==0.5.7
decorator==5.1.1
docker==6.1.3
entrypoints==0.4
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.10.2
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-pubsub==2.18.0
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.48.2
h5py==3.9.0
hyperlink==21.0.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
incremental==22.10.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jsonschema-specifications==2023.6.1
keras==2.13.1
langchain==0.0.245
langsmith==0.0.14
libclang==16.0.6
looker-sdk==23.8.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mendel-development-tool==1.5.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
paramiko==3.0.0
Pillow==9.5.0
ply==3.11
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycairo==1.23.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.15.1
PyGObject==3.44.1
Pympler==1.0.1
PyNaCl==1.5.0
pyrsistent==0.19.3
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rpds-py==0.8.11
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
Shapely==1.8.5.post1
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.49
sqlalchemy-bigquery==1.7.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tensor==0.3.6
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
Twisted==22.10.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
websocket-client==1.6.1
Werkzeug==2.3.6
whisper==1.1.10
wrapt==1.15.0
yarl==1.9.2
zeroconf==0.47.3
zipp==3.15.0
zope.interface==6.0